Allow connection consumers to force a reconnect
Consumers of the library can set timeouts on operations and force a
reconnect in the event of timeouts. This moves bad connections more
quickly into a disconnected state, allowing consumers to remove them
from the connection pool into a new connection has been successfully
established

There is an argument that the semantics of `intentionally_closed`
don't quite fit with this use case - in the force_reconnect method, the
connection was `intentionally_closed` too - a better description might
be `desired_state = connected` or `should_be_connected`. Open to
suggestions